Complete support for #2
If you're using Redis, get TTL just works.

If you're using Memcached, you need to set get_ttl_support on
the cache backend - BEFORE you set any values with a TTL, e.g.

    $cache->backend->get_ttl_support(1);

This is emulated, and a second key is written to Memcached
for every key set with a TTL.